# Backend_Jobaile/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jobaile API",
        "description": "Dokumentasi API untuk aplikasi Jobaile (versi pertama)",
        "contact": {
            "name": "Jobaile Dev Team",
            "email": "support@jobaile.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Server utama"
        }
    ],
    "paths": {
...
        "/api/contracts/{id}": {
            "post": {
                "tags": [
                    "Contract"
                ],
                "summary": "Recruiter membuat kontrak kerja dengan worker",
                "operationId": "25e02c2ca7bd67c5e8e4389ac3875806",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dari worker yang akan dikontrak",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_job",
                                    "start_date",
                                    "end_date",
                                    "terms"
                                ],
                                "properties": {
                                    "id_job": {
                                        "type": "string",
                                        "example": "job123"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-07-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-07-10"
                                    },
                                    "terms": {
                                        "type": "string",
                                        "example": "Kontrak berlaku 10 hari dan dapat diperpanjang."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kontrak berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Kontrak berhasil dibuat."
                                        },
                                        "kontrak": {
                                            "properties": {
                                                "id_worker": {
                                                    "type": "string",
                                                    "example": "worker123"
                                                },
                                                "id_recruiter": {
                                                    "type": "string",
                                                    "example": "recruiter123"
                                                },
                                                "id_job": {
                                                    "type": "string",
                                                    "example": "job123"
                                                },
                                                "start_date": {
                                                    "type": "string",
                                                    "example": "2025-07-01"
                                                },
                                                "end_date": {
                                                    "type": "string",
                                                    "example": "2025-07-10"
                                                },
                                                "status_pay": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "terms": {
                                                    "type": "string",
                                                    "example": "Detail syarat dan ketentuan"
                                                },
                                                "sign_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-06-23T12:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User belum terautentikasi"
                    },
                    "403": {
                        "description": "Email belum diverifikasi atau bukan recruiter"
                    },
                    "404": {
                        "description": "Profil recruiter tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal (format atau data tidak sesuai)"
                    },
                    "500": {
                        "description": "Kesalahan server"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}